+++ appyx-components/experimental/puzzle15/common/src/commonMain/kotlin/com/bumble/appyx/components/experimental/puzzle15/Puzzle15.kt
@@ -17,11 +17,11 @@ class Puzzle15(
     scope: CoroutineScope = CoroutineScope(SupervisorJob() + Dispatchers.Main),
     model: Puzzle15Model = Puzzle15Model(savedStateMap = null),
     motionController: (UiContext) -> MotionController<Tile, Puzzle15Model.State> = {
-        Puzzle15MotionController(
-            it
-        )
+        Puzzle15MotionController(it)
     },
-    gestureFactory: (TransitionBounds) -> GestureFactory<Tile, Puzzle15Model.State> = { bounds -> Puzzle15MotionController.Gestures(bounds) },
+    gestureFactory: (TransitionBounds) -> GestureFactory<Tile, Puzzle15Model.State> = { bounds ->
+        Puzzle15MotionController.Gestures(bounds)
+    },
     animationSpec: AnimationSpec<Float> = spring(),
     animateSettle: Boolean = false,
 ) : BaseAppyxComponent<Tile, Puzzle15Model.State>(