+++ appyx-navigation/common/src/androidMain/kotlin/com/bumble/appyx/navigation/integration/permissionrequester/PermissionRequestResultHandler.kt
@@ -1,0 +1,10 @@
+package com.bumble.appyx.navigation.integration.permissionrequester
+
+interface PermissionRequestResultHandler {
+
+    fun onRequestPermissionsResult(
+        requestCode: Int,
+        permissions: Array<out String>,
+        grantResults: IntArray
+    )
+}