+++ appyx-navigation/common/src/androidMain/kotlin/com/bumble/appyx/navigation/integration/requestcode/RequestCodeDoesntFitInMask.kt
@@ -1,0 +1,3 @@
+package com.bumble.appyx.navigation.integration.requestcode
+
+class RequestCodeDoesntFitInMask(override val message: String?) : RuntimeException(message)