+++ appyx-navigation/common/src/commonMain/kotlin/com/bumble/appyx/navigation/lifecycle/LocalCommonLifecycleOwner.kt
@@ -1,0 +1,7 @@
+package com.bumble.appyx.navigation.lifecycle
+
+import androidx.compose.runtime.ProvidableCompositionLocal
+import androidx.compose.runtime.compositionLocalOf
+
+val LocalCommonLifecycleOwner: ProvidableCompositionLocal<CommonLifecycleOwner?> =
+    compositionLocalOf { null }