+++ appyx-navigation/common/src/commonMain/kotlin/com/bumble/appyx/navigation/node/ViewFactory.kt
@@ -1,0 +1,5 @@
+package com.bumble.appyx.navigation.node
+
+fun interface ViewFactory<out View : NodeView> {
+    operator fun invoke(): View
+}