+++ appyx-navigation/common/src/jsMain/kotlin/com/bumble/appyx/navigation/integration/MainIntegrationPoint.kt
@@ -1,0 +1,15 @@
+package com.bumble.appyx.navigation.integration
+
+import com.bumble.appyx.navigation.integrationpoint.IntegrationPoint
+
+class MainIntegrationPoint : IntegrationPoint() {
+    override val isChangingConfigurations: Boolean
+        get() = false
+
+    override fun onRootFinished() {
+    }
+
+    override fun handleUpNavigation() {
+
+    }
+}