+++ appyx-navigation/src/main/kotlin/com/bumble/appyx/navigation/children/ChildCallback.kt
@@ -1,7 +1,0 @@
-package com.bumble.appyx.navigation.children
-
-import androidx.lifecycle.Lifecycle
-
-typealias ChildrenCallback<T1, T2> = (commonLifecycle: Lifecycle, child1: T1, child2: T2) -> Unit
-
-typealias ChildCallback<T> = (commonLifecycle: Lifecycle, child: T) -> Unit