+++ appyx-navigation/src/main/kotlin/com/bumble/appyx/navigation/node/ViewFactory.kt
@@ -1,3 +1,0 @@
-package com.bumble.appyx.navigation.node
-
-fun interface ViewFactory<out View : NodeView> : () -> View