+++ demos/mkdocs/appyx-interactions/gestures/dragprediction/web/src/jsMain/kotlin/com/bumble/appyx/demos/dragprediction/DragPrediction.kt
@@ -158,7 +158,10 @@ fun Target(
         Box(
             modifier = modifier
                 .size(60.dp)
-                .offset(targetUiState.position.value.offset.x, targetUiState.position.value.offset.y)
+                .offset(
+                    targetUiState.position.value.offset.x,
+                    targetUiState.position.value.offset.y
+                )
                 .scale(targetUiState.scale.value)
                 .rotate(targetUiState.rotationZ.value)
                 .alpha(alpha)